[Recovery] Add buttons for WiFi connection
These buttons do not do anything yet but are displayed on
the bottom of the screen. They will be hooked up in the next CL.

This is the third of several connected CLs for the on-screen
keyboard which uses buttons to show the keyboard.

Bug: b/202857033 b/203107208
Test: None
Change-Id: Ia8f4674450230d32ab82f68add3f026f3a3bb67f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/620163
Reviewed-by: Rob Tsuk <robtsuk@google.com>
Reviewed-by: Mike "mc" Comfoltey <comfoltey@google.com>
Commit-Queue: Kevin Twidle <kpt@google.com>