[media/vnext] Update to fuchsia.audiovideo.Producer
This CL updates the FIDL definition for fuchsia.audiovideo.Producer
to address issues that arise when requesting compression types other
than the default 'native' compression type.

An A/V producer (a demux, basically) can now alter the compression
on produced streams. This required some updates to the FIDL
definition.

1) ProducerStreamInfo, which describes a stream, is now a
table and describes the native format and compression as well as
available alternative compression types.

2) The ProducerStream definition has also been updated to have a
separate SetCompressionType method rather than just taking Connect
parameters that specify compression. This was done because the format
produced after a decoder is added may be different than the original
format and because, in encode/transcode cases, the new compression
may have a parameters blob that needs to be specified in the connect
call.

Test: fidl changes only
Change-Id: Ie1ef80333836a854b1c0976965dc1cf988b9f8dd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/626556
Reviewed-by: Tom Bergan <tombergan@google.com>
API-Review: Dale Sather <dalesat@google.com>
Commit-Queue: Dale Sather <dalesat@google.com>