[ffx] Add flags for TUN/TAP networking
This CL adds the basic flags for running the emulator in Tap network
mode, by adding a tap interface on the default mac address. It makes no
effort to set up the interface, but prints an error if it's not already
configured. It will also print a warning if Tap mode is attempted on
MacOS.

Manual testing shows this is sufficient to enable a user to access the
emulator through `fx shell` on a Linux machine with a Tap interface in
place.

Bug: fxbug.dev/90481
Change-Id: I11fb201bd9911baaa7b796d6b4990968f8868599
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/627923
Fuchsia-Auto-Submit: Daniel Dougherty <doughertyda@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>