[cursor] Add cursor component
This implements a simple cursor component using carnelian and
makes the Flatland scene manager use it to display a cursor.

The cursor component produces the appropriate output for
a cursor be rendering a rive file. The component can be
updated to support different cursor types and animations
in the future.

The Flatland scene manager implement a cursor sprite by
embedding the cursor components view, and controlling its
size and location in the scene to match the location of
the pointer.

Carnelian has been updated to allow clients to disable input.
This is required to not have the cursor view steal all input
until we have a better way for the scene manager to handle
this.

MouseInjectorHandler has been updated to match the gfx mouse
handler by centering the cursor initially and supporting
cursor hiding on touch and immersive mode.

Test: input_pipeline_lib_tests
Change-Id: I27b622b5fa1826f1284bc156f03d07b0806281d2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/613242
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
Reviewed-by: Jaeheon Yi <jaeheon@google.com>
Reviewed-by: Josh Gargus <jjosh@google.com>
Commit-Queue: David Reveman <reveman@google.com>