[fidl][drivers] Move client APIs behind `.buffer()`
This aligns the driver client closer to the regular fidl::WireClient, so
that they both use `.buffer(...)` to pass in an arena, and get back an
interface to make calls.

This is accomplished by creating driver-specific client wrappers,
fdf::WireClient and fdf::WireSharedClient. They are more-or-less copied
from their fidl:: counterparts, but with bits and pieces tweaked to be
more driver specific:

  - Takes fdf_dispatcher_t* instead of async_dispatcher_t*
  - `.buffer(...)` takes a `fdf::Arena&&` instead of any memory resource
  - `client->FooMethod()` is not supported. We don't have an immediate
  need for the managed memory flavor.
  - Some comments are shortened to reference the fidl:: counterpart
  to reduce future matainence burden

TEST: tested that DriverTransport.DISABLED_TwoWayAsync works up to
canceling the channel wait when un-disabled.

Change-Id: I0cd84f05c42e88bbb7730463bc922492a78756b4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/627756
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Benjamin Prosnitz <bprosnitz@google.com>
Commit-Queue: Yifei Teng <yifeit@google.com>