[component_manager][routing] service.rs uses PartialAbsoluteMoniker
The `.collection_component` and `.intermediate_component` fields now
refer to the `PartialAbsoluteMoniker` instead of the `AbsoluteMoniker`
of these respective components. These values are only used for logging,
no change in behavioral logic is expected.

The reason for this change is to hide AbsoluteMonikers from the public
API space.

Bug: 87407
Change-Id: I4932451da5cc89fbad52d9b2c31e953d72d17130
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/628561
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
Commit-Queue: Caslyn Tonelli <caslyn@google.com>