[testrunner] Handle non-fatal errors within Test().
I added a TestTimeout status and a FailReason to the testrunner
TestResult so that we can set those values within the TestResult
returned by Test() and have it only return fatal errors. That way,
main.go doesn't need to check the type of error it gets from Test() to
see if it's fatal or not.

Bug: 49266
Change-Id: Idddef2933ed2f80ced890aaa56ba57355a962b8a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/628089
Reviewed-by: Oliver Newman <olivernewman@google.com>
Commit-Queue: Ina Huh <ihuh@google.com>