[login] add a hosting directory to hold account_data
The flutter runner expects that all top level directories
exist at the time that outgoing is served since it
is using a composed directory. We have to put in one
more level of indirection since account_data does not
exist when out/ is served. This allows us to add the
account_data at a later time after the user has logged in.
BUG: 89690

Change-Id: Ia38b2c1a786ac6020b3bf925fd286c53536df3a9
Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/628505
Reviewed-by: Sanjay Chouksey <sanjayc@google.com>
Commit-Queue: Chase Latta <chaselatta@google.com>