[kernel][vm] Dirty requests for zero pages
Generate ZX_PAGER_VMO_DIRTY page requests for offsets in pager-backed
VMOs that have a zero page marker. For pager-backed VMOs, a zero page
marker represents a clean zero page, so writes to it should allocate a
page and mark it dirty.

If the VMO is created with ZX_VMO_TRAP_DIRTY, a write to a zero page
marker forwards a ZX_PAGER_VMO_DIRTY page request to the user pager,
which the user pager then fulfills with ZX_PAGER_OP_DIRTY. At this
point, the zero page is forked and marked dirty.

Test: core-pager-writeback-test

Bug: 63989
Bug: 88674
Change-Id: I19349abe537e4a1ba0ab92127fc6ff49f9072fe5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/627037
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>