[kernel][vm] Extend dirty requests over lookup range
When generating ZX_PAGER_VMO_DIRTY requests, try to extend the range to
include other clean pages and zero page markers that are also going to
need requests. This is an optimization to reduce the number of distinct
requests that need to be sent overall.

Before returning from LookupPagesLocked, extend the range to cover the
max lookup length, regardless of how many pages were actually looked up.
This is fine, because we are not returning any looked up pages when a
dirty request is generated. We are simply forwarding the request to the
user pager and returning with ZX_ERR_SHOULD_WAIT.

Test: core-pager-writeback-test

Bug: 63989
Bug: 88674
Change-Id: I58aca4537f37fc270a211eb4c8812bcc31015e19
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/627228
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>