[kernel] Add EntropyPool::AddFromDigest
Adds logic to entropy pool to add the digest of input
bytes instead of the bytes directly.

This will be used on phys, so validated hex strings are
added.

The size of the digest is treated as an implementation detail,
so it is returned, so the caller can perform any entropy estimation.

Also made a public config for the include directory.

Test: k ut crypto-entropy-pool
Change-Id: Ic97c906759418dc778ebe43ba4adcde4e2a721c7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/612402
Reviewed-by: Marco Vanotti <mvanotti@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>