[testrunner] Mock ssh.ExitError
Instead of unwrapping an error as an ssh.ExitError to check if the test
command had a retcode that indicates a timeout, unwrap it as an
interface implementing only the subset of ssh.ExitError's methods that
we need to use. This makes it possible to create a fake error type to
test handling of timeouts without having a separate testing-specific
code path, which was previously necessary because it's not possible to
construct an ssh.ExitError in-memory due to its fields being private.

Change-Id: Ibd01358f675d4b9fc3734ed416425522e57deb5d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/629181
Commit-Queue: Oliver Newman <olivernewman@google.com>
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Reviewed-by: Ina Huh <ihuh@google.com>