[tools][fuzz] Retry failed SSH connections from undercoat
In addition to the retry logic added in the previous CL that is specific
to initial instance boot, this CL adds retry behaviour at a lower level
so that it will apply to all SSH commands run against the instance.

The two levels of retry logic remain separated because `start_instance`
needs to handle the edge case where the connection succeeds but the
remote command fails.

Change-Id: I645135213807320a04fbc8d4b6a78af0a72a583b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/627880
Reviewed-by: Aaron Green <aarongreen@google.com>
Commit-Queue: Cameron Finucane <eep@google.com>