[debug_data] Make data processing accessible via fidl
Creates a new 'debug_data_processor' component that contains all the
VMO processing logic. This allows us to call it via fidl from a
component written in Rust. The component is not used yet.
Note that infra scripts currently rely on the location of the debug
data on the target's filesystem for pulling coverage. To support this
case, we allow the caller to pass in a directory the processor will
work in. We can freely change this detail once the scripts are
migrated.

Tests: new unit tests
Bug: 77634

Change-Id: I5bc47edbb6dc2890a2f7a690f83c7cd62c65dfb1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/622793
Reviewed-by: Ankur Mittal <anmittal@google.com>
API-Review: Christopher Johnson <crjohns@google.com>
Commit-Queue: Satsuki Ueno <satsukiu@google.com>