[rust][cf] Remove unused Capability::event method
This CL removes an unused method to unblock the Rust toolchain roll,
since it triggers a dead_code warning in the new toolchain. Although the
method is marked `pub`, the mod is private to the crate and the method
is never used.

Bug: 91243
Change-Id: Id82b91b994533dc9fb57f4fe2e8af783e313b207
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/630101
Fuchsia-Auto-Submit: Mitchell Kember <mkember@google.com>
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
Reviewed-by: Yaneury Fermin <yaneury@google.com>
Commit-Queue: Mitchell Kember <mkember@google.com>