[fidlc] Provide spans for compile_step errors.
Update call sites for ErrOrOperatorOnNonPrimitiveValue,
ErrMismatchedNameTypeAssignment, ErrTooManyTableOrdinals, and
ErrMaxOrdinalNotTable. to use Fail instead of FailNoSpan.

Remove ErrCouldNotResolveIdentifierToType because it appears to be
unreachable and no test cases cover it.

Change ErrResourceMustBeUint32Derived to be enforced when compiling a
resource rather than when resolving bounds for uses of that resource.

Bug: 89213
Test: fx test fidl-compiler
Change-Id: I5f879ce169fcaeae0ae2f7a219f0a6b759fb513c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/629361
Reviewed-by: Mitchell Kember <mkember@google.com>
Commit-Queue: Zachary Stewart <zstewart@google.com>