[fidl][cpp] Natural types for unified bindings.
This generates natural types for unified bindings. They're written from
scratch with the aim of being simple and familiar to C++ programmers,
consistent and hard to misuse.

As part of this we also now generate a common_types.h for types that are
shared between the wire and natural bindings - specifically enums and
bits.

Fixed: 82189

Change-Id: I2f5ff45c6735ea10f572661471d86b4a8bfe186e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/604191
Reviewed-by: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Pascal Perez <pascallouis@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Ian McKellar <ianloic@google.com>