[vulkan] Wire //src/lib/vulkan/tests up to //src:tests.
Tests in //src/lib/vulkan/tests should be wired up to
its parent directory so that it could be a dependency item
of //src:tests and could be tested in infra.

Also this change limits the environment of "compact-image-tests"
to devices with ARM GPU since only ARM driver supports the
memory control extension which is required by the compact image
layer.

TEST: CQ
Change-Id: Ib41cda1d7805a67a2c5f0209f9dea461c1a09177
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/628802
Reviewed-by: John Bauman <jbauman@google.com>
Commit-Queue: Yilong Li <liyl@google.com>