[cobalt] Disable aggregation backfill in cobalt_testapp.
The backfill doesn't find any metadata for the days in the past,
which causes the MetadataBuilder to log a lot of errors. This
partially depends on fxr/623241 which makes the backfill days
flag work for Cobalt 1.1 as well, though submitting this first
won't break anything.

Bug: 89960
Change-Id: I7f0d900e363bd7d3831c137cba3d60f784891da6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/622800
Reviewed-by: Zach Bush <zmbush@google.com>
Commit-Queue: Cameron Dale <camrdale@google.com>