[llcpp] Remove Fuchsia-specific code from transport.h
The main thing requiring __Fuchsia__ in transport.h previously was a
reference to the async_dispatcher_t. However, this type is only needed
in transport.h to forward to other things and can be redeclared.
In the future, it may be desirable to migrate to something less
fuchsia-dependent than an async_dispatcher_t but given that it will look
identical to an async_dispatcher_t, it is easy to stick with the
existing type for now.

Change-Id: If74ae2b16686e38795b6d9404fb2b935cf8b8794
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/630964
Fuchsia-Auto-Submit: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>