[pkg-resolver] PackageFetcher type alias -> wrapper struct
This change is entirely a no-op in hopes of making the package fetch
queue a bit easier to follow. The idea is if PackageFetcher is a wrapper
type it'll be easier to control-click your way to find the inner queue
implementation.

Because control-clicking already would take you to the type alias, which
was right next to the construction function, it's only a minor
improvement, but it still seems worthwhile.

A pleasant side-effect is the doc comments get a bit nicer to read,
since they're a lot less abstract (since the queue's processing
behavior is known rather than being arbitrary)

Change-Id: I3db39a57ad5f93839e26f0d76e2699fab3cf5f92
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/628587
Reviewed-by: Kevin Wells <kevinwells@google.com>
Reviewed-by: John Wittrock <wittrock@google.com>
Fuchsia-Auto-Submit: Dan Johnson <computerdruid@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>