[Bootsvc] Stop creating bootfs when flag passed.
This change stops creating the bootfs service when --host_bootfs
is passed as an argument to bootsvc.next, as component manager will
create a Rust bootfs VFS instead.

The behaviour when --host_bootfs is not passed should be
unchanged, and most of the changes are in support of optionally creating
a bootfs service. A number of key things to note:

1) Parsing CMDLINE arguments has been moved earlier, because we need to
know if --host_bootfs was passed.
2) Args from config/devmgr come directly from the image when bootfs
isn't created.
3) The bootfs loader service (needed to load component manager) gets
dynamic libs directly from the image when bootfs isn't created.
4) bin/component_manager comes directly from the image when bootfs isn't
created.

Once we're confident in the new Rust bootfs VFS, it will be removed from
bootsvc, and once bootsvc isn't doing anything, userboot will load
component manager directly.

Bug: 44784

Testing:
-Existing unit and integration tests (poor coverage of the experimental
bootfs for now, but good coverage to prevent regressions)
-Manual testing

Change-Id: Ib1c53b7f6dd15e10003c191111191f2278422f81
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/625803
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Commit-Queue: David Hastings <dahastin@google.com>