[audio][tools] SigGen specify PTS for first packet
The CLI utility signal_generator can now specify the PTS value
for the first packet it submits to the AudioRenderer. The cmdline
flag '--packet-pts' indicates that timestamps should be applied
to each packet; it is now extended to accept a value that will be
used as the PTS for the first packet submitted. This flexibility
is useful when interactively exploring cases where the packet
stream starts after (or before) the renderer starts playback.

fx shell signal_generator --rate=48000    --pts-unit=48000/1 \
                          --media-start=0 --packet-pts=48000

Change-Id: Id53cb7543a9481c4007f9c901fb0994135748a96
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/627877
Reviewed-by: Tom Bergan <tombergan@google.com>
Reviewed-by: Alper Gungormusler <alperg@google.com>
Commit-Queue: Martin Puryear <mpuryear@google.com>