[starnix] Set process debug address
This change sets the process' debug address to the one present in the
DT_DEBUG entry, when such an address is set by the linker.

If the process' debug address is set to `ZX_PROP_PROCESS_DEBUG_ADDR`
starnix will issue a software interrupt when the property is set to
inform the debugger.

Change-Id: Ice91d6fc96d8f63fbc95614dfd6b677bf23197fd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/620722
Fuchsia-Auto-Submit: Kevin Lindkvist <lindkvist@google.com>
Reviewed-by: Adam Lesinski <adamlesinski@google.com>
Reviewed-by: Dangyi Liu <dangyi@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>