[pwauth] Return more specific error when derive_key fails
When an account is created with an empty password, then it is bound to
the "NullKey" account metadata variant.  If you attempt to unlock that
account with a non-empty password, `derive_key` fails, since otherwise
there's nowhere else for us to check the password, but the FIDL error we
return for that failure was Unknown, which is confusing.  It should
instead be AuthenticationFailed.

Fixed: 91125
Change-Id: I434d84ec339695d16517db07a2144c0595cd2bb5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/631083
Reviewed-by: Jody Sankey <jsankey@google.com>
Commit-Queue: Drew Fisher <zarvox@google.com>