[pretty] Add support for pretty::ParseSizeBytes
pretty::ParseSizeBytes does the reverse, to some extent, of
format_string. ParseSizeBytes will take a formatted string
and will return the number of bytes the string tries to encode.

e.g.
64k = 64 * 1024
64.25k = 64 * 1024 + 256

As commented in the source code, not to remove support for larger
units, and to be able to convert strings formatted from the library
manual software division and multiplication was added.

This is necessary sin a perfectly valid and representable size
10.4492187500t (and above) could cause the simplest approach to
overflow. To work around this, we add on element of the mantisa
at a time, and keep track of the accumulated carries from each
scaled element. Essentially this is reproduceable with any long
mantisa.

Bug: 76151
Test: fx test pretty-test-test
Change-Id: Ifcf14431a29b949a2ff6295cbc25e6bac4ddafde
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/614122
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>