[ftl] Set up some basic unit testing for the ftl.
Sets up the ability to reach inside the Volume and underlying ram
storage to be able to inspect and manipulate the current state in order
to introduce desired failure modes.

Tests added here include a partial page write example as it is a
previously experienced issue, as well as a couple of behaviours that I
was fairly confident in, relying upon, but not really certain about.

Bug: 38433

Change-Id: I527a0c4c7a3d842062211f0996c052f3c310a73d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/628091
Reviewed-by: Brandon Castellano <bcastell@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>