[testrunner] Don't write test data into the source tree
Previously the testrunner main_test would inadvertently write out files
relative to main_test.go when run with `go test` because the fake test
output dir was set to an empty string.

And enforce that the outdir passed to `CreateTestOutputs` is not an
empty string, so it's safe for the TestOutputs to assume that OutDir is
set.

Change-Id: I0a1bc57103b3687ce3d42412335f2f2b2acd8a65
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/631701
Reviewed-by: Ina Huh <ihuh@google.com>
Reviewed-by: Gulfem Savrun Yeniceri <gulfem@google.com>
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>