[system-update-checker] Fallback update channel to ""
Previously, the system-update-checker upon startup would try to discover
the update channel by:

1. Check if vbmeta contains an `ota_configs` channel.
2. Otherwise, use the rewrite engine to rewrite
   fuchsia-pkg://fuchsia.com/update/0, and use the rewritten URL's
   hostname as the update channel name.

The downside of this approach is that it's not stable. If the rewrite
rules for `fuchsia.com` are changed and the device is rewritten, the
device will think it's now on a different channel.

This patch changes the behavior to use "" as the current and target
channel if one is not discovered in vbmeta. This better reflects the
fact that we can't reliably know the origin of the current build.

Fixed: 91158
Change-Id: I0a02c3e628b86aa6c909a400aab1d9e9cb36f079
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/627872
Reviewed-by: Dan Johnson <computerdruid@google.com>
Commit-Queue: Erick Tryzelaar <etryzelaar@google.com>