[fidl][cpp][natural] Reorganize struct template
Two changes here:

The name of the type that holds the actual data is renamed from Storage
to Storage_ and made private. This will help prevent conflicts with
nested types.

Most of the private parts of the class have been moved to after the
public parts. This is to follow C++ coding guidelines. The only private
declaration that's above the public ones is a forward declaration for
the Storage_ type since it's referred to in the public API.

Change-Id: Idbe197c6c18c2ebd3f336968fdaa8b1309062342
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/631446
Fuchsia-Auto-Submit: Ian McKellar <ianloic@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Ian McKellar <ianloic@google.com>