[Bootfs] Add a Rust system test.
This adds a system test that briefly checks that the expected directories
exist, and that files in those directories have the correct permissions.
This is based off of (and will replace) the C++ integration test paired
with bootsvc.

Please note that this test adds the '/boot' directory capability to
system tests, which are allowed to be non-hermetic. If we don't want to
extend this capability to all system tests we'll want to add a static
test child instead.

Bug: 44784

Testing:
This is a test.

Change-Id: If9ed218cc3dd0ad2e9e47a765c0e9193003a5dab
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/626182
Reviewed-by: Gary Bressler <geb@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Commit-Queue: David Hastings <dahastin@google.com>