[lowpan][openthread_plat]: Remove config requirement for ResetConnection
In commit `7242f6d6949a9c`, the method `ResetConnection()` was added
to `SpinelFidlInterface`, but the addition was gated on the presence
of the config macro `OPENTHREAD_SPINEL_CONFIG_RESET_CONNECTION`.

It's unclear what the original intent of gating the existence of that
method was, and it now looks to be an oversight. Since it is causing
problems with updating to more recent versions of OpenThread, this
commit removes the macros so that this method is always defined.

Change-Id: Ib729629e3c5cc2493deb1cc0ca7ac5de607b0420


Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/632388
Reviewed-by: Jiaming Wang <jiamingw@google.com>
Commit-Queue: Robert Quattlebaum <rquattle@google.com>