[thermal][test] Use unique stamp files in tests
The new test, validator_bin_test (added in fxrev.dev/620388), appears to
be flaky. Occasionally, one of the test cases will fail because it found
a stamp file when it didn't expect to, or vice versa. This can be
reproduced with the following command (especially running two instances
in separate terminals) which runs the test in a loop until a failure is
encountered: `while out/core.nelson-release/host_x64/validator_bin_test;
do : ; done`

The issue is that both test cases (test_invalid_thermal_config and
test_valid_thermal_config) use the same test-only stamp file path. This
is a problem because one test case writes the stamp file, and the other
test case is verifying the file does not exist. It can happen that one
test case is seeing changes to the stamp file that are caused by the
other test case running simultaneously.

Bug: fxbug.dev/90157
Multiply: validator_bin_test
Change-Id: Ibf6f879925c8965858947a5f6afc115f8cabcf59
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/632092
Reviewed-by: Onath Dillinger <claridge@google.com>
Reviewed-by: Didi She <didis@google.com>
Commit-Queue: Patrick Shickel <pshickel@google.com>