[thermal] Create thermal_config GN template
Creates a new GN template, thermal_config, that wraps two underlying
operations: 1) perform build-time validation of the provided thermal
configuration JSON file, and 2) create a `resource` target with the
valid JSON file to allow it to be included in a package.

This template is intended to be used once per board, passing in a
board-specific thermal configuration file, like so:

  thermal_config("astro") {
    config_file = "astro.json5"
  }

Users of this template will be added in subsequent CLs.

Bug: fxbug.dev/90157
Test: fx test //src/power/power-manager/thermal_config
Change-Id: I06e818a05d1ea9cbfa52a7d7dd9262c7ef4d898c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/621025
Reviewed-by: Onath Dillinger <claridge@google.com>
Reviewed-by: Didi She <didis@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Fuchsia-Auto-Submit: Patrick Shickel <pshickel@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>