[ffx] Add user mode networking to ffx emu
The ffx emu plugin has been operating without user-mode networking for a
while, as we refactored the tool from the ground up. This CL adds it
back.

User-mode networking depends on SLiRP, which operates by defining a
fixed set of port mappings between the host and guest systems. The set
of possible ports are defined in the device_spec, associated with a
user-readable name (e.g. "ssh" is guest port 22). The host ports mapped
to those guest ports are then defined by the user at runtime on the
command line. Any ports defined in the device_spec that are not also
specified on the command line are assigned arbitrary ports dynamically
at runtime using the ffx::port_picker library.

The former "--tuntap" command-line flag has been replaced with the
"--net" flag, which can be "none", "user", "tap", or "auto".

The setup args template is then modified to populate the Qemu
command-line according to these mappings only when using the "--net
user" mode. The rest of the networking flags are unchanged (except to
recognize the modified flag values).

Bug: fxbug.dev/88327
Change-Id: I4cc7ecd19752c6a53ceba79743227457340d213c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/631087
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Dave Schuyler <dschuyler@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>