[ffx] Remove env flag from ffx emu
The ffx emu tool relies on a Product Bundle Manifest for the majority of
its configuration. While some of the fields in the manifest are
configurable through command lines, we wish to minimize the number of
these flags to keep the command line simple. Complicated configuration
should happen directly in the virtual device specification or the setup
flags template.

This CL removes the env flag from the emu tool. This flag was a holdover
from the previous version of the tool, was completely unused in the
current version, and is not the way we want users to modify the runtime
environment anyway.

Change-Id: Idfee48e2a337dc32de202cdce227b189b28a19a4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/632493
Fuchsia-Auto-Submit: Daniel Dougherty <doughertyda@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>