[tools][fuzz] Fix undercoat bug resulting in truncated QEMU log files
Previously, in order to detect a successful boot, undercoat relied on
monitoring the system's serial output via QEMU's stdout, looking for a
specific keyword. After detection, it then detaches from the subprocess.
Additionally, QEMU was also configured to copy its serial output to a
file.

When an unrecoverable error occurs in ClusterFuzz, the system debug logs
are fetched from the QEMU log file to aid in debugging. However, it was
noticed that these were being truncated at approximately 64KB.

The root cause was that, when QEMU is told to output serial to both
stdout and a file, a block on a write to one of those sinks blocks the
writes to the other as well. In this case, undercoat was only reading
the stdout until the boot completion keyword was detected, causing
writes after that to be buffered and then eventually block. Simply
continuing to read and discard from the pipe would also not be
sufficient, because the read end of the pipe is closed after the
`start_instance` command completes and undercoat exits.

To work around this, this CL switches QEMU to write solely to the log
file, which undercoat now watches during the boot process instead.

Tests are also added to verify the handling of logs of greater size than
default pipe buffer sizes, although the coverage is not perfect because
even in the end-to-end tests we do not invoke individual undercoat
commands as separate subprocesses.

Change-Id: Ic190adcfbfc6d31464ae95b4969c2005c2b90f2d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/632095
Reviewed-by: Aaron Green <aarongreen@google.com>
Commit-Queue: Cameron Finucane <eep@google.com>