[memfs] Change memfs C api implementation.
Implements the memfs C API in terms of the new C++ "Setup" object.

While this does not change the C API, it has an important semantic
change in that memfs_free_filesystem() will not remove the namespace
entries created my memfs_install_at() before shutdown. I believe failure
to do this is a bug so this change should not break any callers.

This removes all in-tree uses of memfs_uninstall_unsafe(). I believe
this function was working around the fact that the namespace entries
were not uninstalled and that memfs was not being shut down properly on
its thread.

Bug: 91477
Change-Id: Icd21824e27768aea7d71311012f6472e61f2e7a0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/631642
Reviewed-by: Chris Tam <godtamit@google.com>
Reviewed-by: Jeremy Manson <jeremymanson@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>