[fidl_driver] More LLCPP+Driver tests + fixes
This CL adds a broader set of tests for LLCPP+Drivers. These tests pass
when ASAN is disabled, but fail due to a use after free in teardown so
they are disabled. Still, it is useful to check them in to ensure the
code doesn't become stale.

Change-Id: If1e0558761c2c198d1a68e03668e0ee71813e3f0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/632203
Reviewed-by: Yifei Teng <yifeit@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>