[ffx] Add list filters for ancestors+descendants.
Introduces three new `--only` filters:

- `--only ancestors:foo`: Lists ancestors of a component named `foo`.
- `--only descendants:foo`: Lists descendants of a component named `foo`.
- `--only relatives:foo`: Lists both ancestors and descendants of `foo`.

This is not super useful for the `ffx component list` command, but
it is useful for the `ffx component graph` command, where it can be used
to significantly reduce the size of the graph for illustration.

The implementation accumulates the names of the ancestors for each
component in Component::parse.

Change-Id: I43c206e672942f9e23b46d5ae5b17b9b544322bd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/625602
Reviewed-by: Xyan Bhatnagar <xbhatnag@google.com>
Reviewed-by: Gary Boone <gboone@google.com>
Reviewed-by: Fady Samuel <fsamuel@google.com>
Commit-Queue: Alexander Biggs <akbiggs@google.com>