[fbl][wavl] Implement WAVLTreeBestNodeObserver
Implement WAVLTreeBestNodeObserver, a helper Observer trait struct
which does most of the work required to maintain a secondary "best
value" invariant for an augmented binary tree.

See the comment in
zircon/system/ulib/fbl/include/fbl/wavl_tree_best_node_observer.h, but
this basically lets users define something like "maximum value of X
for a collection of objects" using a Trait struct, and then inject an
observer into a tree implementation such that the "maximum value of X"
is maintained automatically for every subtree in the tree.  This
allows users to easily compute the "maximum value of X" across either
all of the nodes in the tree, or for any partition of the tree's
primary key (eg, "what is the maximum value of X for all of the nodes
with a primary key > Y?").

This augmented tree form is currently used in the kernel's scheduler's
run-queues, and is about to be used in the data structure used by
kernel for wait queues as well (although, with slightly different
semantics).  Subsequent CLs will update the run-queues to use this
adapter, and to introduce the new wait queue structure.

Tests: new unit tests.
Change-Id: I5e30fd19633b1702dca4244b90bba31747a8ddc1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/613637
Reviewed-by: Corey Tabaka <eieio@google.com>
Commit-Queue: John Grossman <johngro@google.com>