[wlan] Rename MacRole to WlanMacRole
Add the Wlan preface to MacRole so that it's more obvious in C
bindings that the MAC role is relative to WLAN and not another
protocol. In code, the difference in C bindings looks like
wlan_mac_role_t instead of mac_role_t. Even in other bindings the
distinction is nice.

Bug: 91601
Test: None. Rename only.
Change-Id: I7b5b8485752170017551edbbae5b315c7993e9a4