[audio_core][test] Tests for gain-ramp FIDL API
Add gain-control-test cases to cover error modes for the FIDL
API GainControl::SetGainWithRamp.

Tests verify that if the ramp's target gain_db is out of range
(above MAX_GAIN_DB, below MUTED_GAIN_DB, or is NAN), the
GainControl and AudioRenderer are both disconnected.

Other cases verify that if the ramp's duration is zero, or is
negative(!), the ramp call is redirected to a SetGain call.
This was already effectively occurring within the Gain class,
but new code in AudioRenderer makes this explicit and in turn
enables the Gain class to be more strict (subsequent CL).

fx test gain-control-test -- --gtest_filter=GainControlRampTest.*

Change-Id: I9c89878098c69877792821540a8990f2ede42e55
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/627878
Reviewed-by: Tom Bergan <tombergan@google.com>
Reviewed-by: Alper Gungormusler <alperg@google.com>
Commit-Queue: Martin Puryear <mpuryear@google.com>