[subprocess] Return ctx.Err() if context cancelled
If the context is cancelled while a command is running, we should return
the context error after cleaning up instead of returning the ExitError
returned by Cmd.Wait() after it gets SIGTERMed. This makes it much
easier for the caller to determine if the subprocess failed to due a
context cancellation or because of an internal failure within the
subprocess.

Notably, this will make it easier for testrunner to determine if a
subprocess test fails due to a timeout or due to an internal failure,
and classify the test failure correctly as a timeout or generic failure.

I also simplified the subprocess runner code in a couple ways:
- Move the cleanup logic into the main thread and run `Cmd.Wait()` in a
  goroutine instead of the reverse. This means less state that has to be
  passed around between functions.
- Because the cleanup logic is now synchronous in relation to starting
  the command, we don't run the cleanup logic unless the command has
  definitely been started, so we don't need to use a mutex to safely
  check if the process has been started.

I tried to avoid introducing any changes to the behavior of how the
runner starts and cleans up subprocesses; the only functional difference
should be the return value of Run().

Bug: 49266
Change-Id: I3e96b5fcd1dd87ec83fe4b092416856613e91d18
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/632265
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Reviewed-by: Ina Huh <ihuh@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>